src/requirements.txt
oss2==2.13.0
pillow==8.0.1